src/test/java/com/cbfacademy/apiassessment/service/FoodItemServiceImplTest.java
package com.cbfacademy.apiassessment.service;

import static org.mockito.Mockito.*;
import static org.junit.jupiter.api.Assertions.*;
import org.junit.jupiter.api.Test;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.springframework.boot.test.context.SpringBootTest;
import com.cbfacademy.apiassessment.repository.FoodItemRepository;
import java.util.Optional;

@SpringBootTest //test supported by Spring Boot
public class FoodItemServiceImplTest {

    @Mock //Mock implementation of the repository
    private FoodItemRepository foodItemRepository;

    @InjectMocks  //Creates an instance of the service 
    private FoodItemServiceImpl foodItemService;

    @Test //Single functionality test
    void whenSaveFoodItem_thenReturnFoodItem() {
        // Arrange
        FoodItem inputFoodItem = new FoodItem(/* parameters */);
        FoodItem savedFoodItem = new FoodItem(/* parameters */);
        when(foodItemRepository.save(any(FoodItem.class))).thenReturn(savedFoodItem);

        // Act
        FoodItem result = foodItemService.saveFoodItem(inputFoodItem);

        // Assert
        assertNotNull(result);
        assertEquals(savedFoodItem, result);
        // Additional assertions as needed
    }

    // Additional tests for retrieve, update, and delete operations
}
